Now I added NASM generation, BUT there is only RETURN and ASSIGN operators. Now I will continue adding features in NASM, then I will do the same in binary code, and then I will add ELF-64

+++ src/Tokens/tokens.h
@@ -50,7 +50,7 @@ struct Token_t
     union NodeType
     {
         int   key_w;
-        double  num;
+        int64_t num;
         char     id[WORD_MAX_LEN + 1] = {};
     } arg;
 };
